example: move exception example to example.common
