+++ src/main/java/com/spring/practice/rest/examples/common/CustomException.java
@@ -1,4 +1,4 @@
-package com.spring.practice.rest.common.exceptions;
+package com.spring.practice.rest.examples.common;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
