+++ src/main/java/com/spring/practice/rest/examples/controller/ResponseStatusController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 
-import com.spring.practice.rest.common.exceptions.CustomException;
+import com.spring.practice.rest.examples.common.CustomException;
 
 import lombok.extern.slf4j.Slf4j;
